Remove `EXPO_METRO_NO_MAIN_FIELD_OVERRIDE`
